refactor(server): create web_app package
move all web-app related files to this package
