+++ server/database.py
@@ -11,10 +11,4 @@ Base.query = db_session.query_property()
 
 
 def init_db():
-    # import all modules here that might define models so that
-    # they will be registered properly on the metadata.  Otherwise
-    # you will have to import them first before calling init_db()
-
-    from models import ContentsCpuUsage
-
     Base.metadata.create_all(bind=engine)
